Move linking into a dedicated module
