+++ src/lib/common/webnative.ts
@@ -1,5 +1,4 @@
 import * as webnative from 'webnative'
-import type { account } from 'webnative'
 import { setup } from 'webnative'
 
 import { asyncDebounce } from '$lib/common/utils'
@@ -112,16 +111,4 @@ export const loadAccount = async (username: string): Promise<void> => {
     username,
     authed: true
   }))
-}
-
-export const createAccountLinkingConsumer = async (
-  username: string
-): Promise<account.AccountLinkingConsumer> => {
-  return await webnative.account.createConsumer({ username })
-}
-
-export const createAccountLinkingProducer = async (
-  username: string
-): Promise<account.AccountLinkingProducer> => {
-  return await webnative.account.createProducer({ username })
 }
