+++ src/stores.ts
@@ -1,10 +1,9 @@
 import { writable } from 'svelte/store'
 import type { Writable } from 'svelte/store'
-import type { account } from 'webnative'
 import type FileSystem from 'webnative/fs/index'
 
-
 import { loadTheme } from '$lib/theme'
+import type { ConsumerState, ProducerState } from '$lib/auth/linking'
 import type { Device } from '$lib/device'
 import type { Session } from '$lib/session'
 import type { Theme } from '$lib/theme'
@@ -22,15 +21,10 @@ export const filesystemStore: Writable<FileSystem | null> = writable(null)
 
 export const deviceStore: Writable<Device> = writable({ isMobile: true })
 
-export const accountLinkingConsumerStore: Writable<account.AccountLinkingConsumer> = writable(
+export const accountLinkingConsumerStore: Writable<ConsumerState> = writable(
   null
 )
 
-export const accountLinkingProducerStore: Writable<{
-  accountLinkingProducer: account.AccountLinkingProducer
-  pin: number[]
-  confirmPin: () => void
-  rejectPin: () => void
- }> = writable(
+export const accountLinkingProducerStore: Writable<ProducerState> = writable(
   null
 )
